+++ src/api.rs
@@ -1,6 +1,6 @@
 pub mod auth;
 pub mod database;
-pub mod kv;
+pub mod kv1;
 pub mod kv2;
 pub mod pki;
 pub mod ssh;
